# contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Identity Onboarding Lab — Public API
  version: 1.0.0
  description: Public HTTP API contract for Sri Lankan investment customer onboarding and reviewer staff operations.
paths:
...
  /applications:
    post:
      summary: Create resumable onboarding draft
      operationId: createApplication
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
      responses:
        '201':
          description: Application draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
  /applications/{id}:
    get:
      summary: Fetch application state and next action
      operationId: getApplication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
  /applications/{id}/details:
    patch:
      summary: Update applicant customer details with optimistic version guard
      operationId: updateDetails
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion, fullName, dateOfBirth, addressLine]
              properties:
                expectedVersion:
                  type: integer
                fullName:
                  type: string
                dateOfBirth:
                  type: string
                  format: date
                addressLine:
                  type: string
                employmentStatus:
                  type: string
                sourceOfFunds:
                  type: string
                taxResidency:
                  type: string
                pepDeclaration:
                  type: boolean
      responses:
        '200':
          description: Details updated
        '409':
          description: Stale version conflict
  /applications/{id}/consents:
    post:
      summary: Record acceptance of versioned consent
      operationId: recordConsent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [consentType, version, accepted]
              properties:
                consentType:
                  type: string
                  enum: [IDENTITY_VERIFICATION, OPTIONAL_COMMUNICATION]
                version:
                  type: string
                accepted:
                  type: boolean
      responses:
        '200':
          description: Consent recorded
...
  /applications/{id}/submit:
    post:
      summary: Submit application for final policy decision evaluation
      operationId: submitApplication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion]
              properties:
                expectedVersion:
                  type: integer
      responses:
        '200':
          description: Application evaluated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
  /applications/{id}/withdraw:
    post:
      summary: Applicant withdraws application
      operationId: withdrawApplication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Application withdrawn
...
  /staff/applications/{id}:
    get:
      summary: Reviewer detail view with masked PII and extracted differences
      operationId: getStaffApplicationDetail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review case detail
  /staff/applications/{id}/decision:
    post:
      summary: Reviewer records manual decision with mandatory reason
      operationId: recordStaffDecision
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion, decision, reason]
              properties:
                expectedVersion:
                  type: integer
                decision:
                  type: string
                  enum: [APPROVED, REJECTED, RECAPTURE_REQUIRED]
                reason:
                  type: string
      responses:
        '200':
          description: Decision updated
...
components:
  schemas:
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lifecycle:
          type: string
          enum: [DRAFT, PROCESSING, READY_TO_SUBMIT, SUBMITTED, UNDER_REVIEW, COMPLETED, WITHDRAWN, REJECTED]
        decision:
          type: string
          enum: [PENDING, RECAPTURE_REQUIRED, REVIEW_REQUIRED, DEMO_CHECKS_PASSED, APPROVED, REJECTED]
        operatingMode:
          type: string
          enum: [LOCAL_DEMO, INSTITUTION]
        version:
          type: integer
        currentDocumentGeneration:
          type: integer
        currentBiometricGeneration:
          type: integer
        submittedAt:
          type: string
          format: date-time
          nullable: true
...